requirements.txt
gradio
langchain
langchain_community
langchain_core
pandas
pyprojroot
python-dotenv
PyYAML
SQLAlchemy
langchain_google_genai